+++ smart-campus-service-impl/pom.xml
@@ -42,12 +42,10 @@
 		<dependency>
 			<groupId>com.fasterxml.jackson.core</groupId>
 			<artifactId>jackson-core</artifactId>
-			<version>2.8.6</version>
 		</dependency>
 		<dependency>
 			<groupId>com.fasterxml.jackson.core</groupId>
 			<artifactId>jackson-databind</artifactId>
-			<version>2.8.6</version>
 		</dependency>
 		<dependency>
 			<groupId>hu.unideb</groupId>
